requirements.txt
Flask
python-docx
reportlab
